finish hibernate unit test

+++ src/main/java/com/ascending/mingqian/repository/UserDaoImpl.java
@@ -7,6 +7,7 @@ import com.ascending.mingqian.util.HibernateUtil;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
 import org.hibernate.query.Query;
+import org.hibernate.query.criteria.internal.expression.NullLiteralExpression;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -39,44 +40,39 @@ public class UserDaoImpl implements UserDao{
 
     @Override
     public boolean delete(String userName){
-//        sql1 = "delete from record where record.account_id in " +
-//                "(select id from account where account.users_id = "+id+")";
-//        sql2 = "delete from account where account.users_id = "+id;
-//        String hql1 = "DELETE Record where account_id in" +
-//                "(FROM account AS account where users_id = " +
-//                "(FROM User where name = :userName1))";
-        User u = getUserByName(userName);
-       // List<Account> a = accountDao.getAccountByUserId(u.getId());
 
-//        for(Users u : users){
-//            System.out.println(u.getId()+" "+u.getName()+" "+u.getPassword());
-//
-//        }
-        Account a = accountDao.getAccountByUserId(u.getId());
-        System.out.println(a.toString());
-        Record r = recordDao.getRecordByAccountId(a.getId());
-        System.out.println(r.toString());
-        recordDao.delete(r.getId());
-        accountDao.delete(u.getId());
 
-        String hql = "DELETE User where name = :userName1";
-        int deletedCount = 0;
-        Transaction transaction = null;
-        try(Session session = HibernateUtil.getSessionFactory().openSession()){
+            User u = getUserByName(userName);
+            if(u != null) {
 
-            Query<User> query = session.createQuery(hql);
-            query.setParameter("userName1",userName);
+                List<Account> accounts = accountDao.getAccountByUserId(u.getId());
+                for (Account a : accounts) {
+                    List<Record> records = recordDao.getRecordByAccountId(a.getId());
 
-            transaction = session.beginTransaction();
-            deletedCount = query.executeUpdate();
-            transaction.commit();
-        }
-        catch (Exception e){
-            if (transaction != null)transaction.rollback();
-            logger.error(e.getMessage());
-        }
-        logger.debug(String.format("The user %s was deleted",userName));
-        return deletedCount >= 1 ? true : false;
+                    for (Record r : records) {
+                        recordDao.delete(r.getId());
+                    }
+                    accountDao.delete(a.getId());
+                }
+            }
+
+                String hql = "DELETE User where name = :userName1";
+                int deletedCount = 0;
+                Transaction transaction = null;
+                try (Session session = HibernateUtil.getSessionFactory().openSession()) {
+
+                    Query<User> query = session.createQuery(hql);
+                    query.setParameter("userName1", userName);
+
+                    transaction = session.beginTransaction();
+                    deletedCount = query.executeUpdate();
+                    transaction.commit();
+                } catch (Exception e) {
+                    if (transaction != null) transaction.rollback();
+                    logger.error(e.getMessage());
+                }
+                logger.debug(String.format("The user %s was deleted", userName));
+                return deletedCount >= 1 ? true : false;
 
 
     }
@@ -123,13 +119,35 @@ public class UserDaoImpl implements UserDao{
             query.setParameter("name",userName.toLowerCase());
 
             User user = query.uniqueResult();
-            logger.debug(user.toString());
+            if(user != null) {
+                logger.debug(user.toString());
+            }
+                return user;
 
-            return user;
+
 
         }
 
     }
+    public User getUserById(Long id){
+        if(id == null) return null;
 
+        String hql = "FROM User as user where user.id = :id";
+
+        try(Session session = HibernateUtil.getSessionFactory().openSession()){
+            Query<User> query = session.createQuery(hql);
+            query.setParameter("id",id);
+
+            User user = query.uniqueResult();
+            if(user != null) {
+                logger.debug(user.toString());
+            }
+            return user;
+
+
+
+        }
+
+    }
 
 }
